Cudd_CountMinterm: revert to 2.5.1 behaviour
In cudd 3.0.0, the computation in Cudd_CountMinterm has been changed. Instead of starting with 2^numvars as the number of minterms when there are not significant variables, it now uses 2^(numvars-bias), which results in extremely small values. This is problematic in PRISM as it pollutes the constant table with those values and can interact badly with the constant collapsing / truncation as currently implemented.

We therefore revert here back to the 2.5.1 behaviour, at the cost of less precision in the presence of a high number of variables.

To experiment with the new behaviour, compile with -DCUDD_COUNT_MINTERM_3_0_0.